Export Db API for integration tests
This is a part of #3.

This patch exports API for integration tests by adding pub visibility to
them. It also adds needed interface for manipulating types and
variables.

Additionally it adds a possibility to provide custom db actor - to be
able to mock ScyllaDB functionality, so there is some small redesign in
main.rs and lib.rs: main creates new db actor before running services.