Introduce vector of i64s as a base for multicolumn primary keys
This is a part of #23.

This is first step in introduction multicolumn primary keys. This change
introduces a vector storage of i64s for a primary key as a collection of
several columns. Currently it will use only first item in this vector
(the size will be always 1) - the one which is retrieved from
IndexMetadata. In the next patch there will be support more dimensions.
It changes HTTP API according to this change also.