Refactor reading embeddings as a channel
This is a part of #54.

This patch creates a channel of embeddings from db index while
constructing its owner, instead of stream returned from API. There
should be the only one channel/stream in the lifetime of db index, so it
is better to create it during construction. A mpsc channel is better
solution for concurrent scan ranges introduces in the next patch.

There should be also a limitation of concurrent requests by the index actor, so
there is implemented a semaphore to limit that.