Remove temporary storage of index parameters
This is a part of #54.

The old workaround for storing metadata of an index in specific table in
the ScyllaDB has been removed. This patch removes getting parameters of
an index from specific table and replaces them with default values.
Setting index parameters from the custom index in CREATE INDEX will be a
part of #57.