Remove index element count storage from a db
This is a part of #54.

As now a vector-store has HTTP API for retrieval of a size of an index,
the old workoaround for this functionality using a specific table/field
could be removed. This patch removes support for a counter of an index
size in the ScyllaDB, it removes not needed actor API. It also removes
housekeeping functionality from an usearch index actor as it was
responsible only for updating the counter in the ScyllaDB.