gh-24: read timestamps for embeddings from a db
The ScyllaDB uses cells timestamps to resolve merging data problem -
only a modification with the latest timestamp is stored in a db. The
vector-store needs to deal with the same problem - data from the
ScyllaDB will flow concurrently, so there is a need for coordinating
them with timestamps - data with the latest timestamp should be stored
in the index.

This patch implements using same timestamps for embeddings as the
ScyllaDB. It uses `SELECT id, embedding, writetime(embedding)` for a
retrieval of embeddings with timestamps. Timestamps in the ScyllaDB
writetime are in microseconds since a unix epoch. Internally this patch
introduces a newtype Timestamp with time::OffsetDateTime.

As support for Deserializing Vector type within Row has been implemented
in a rust driver there is a removal of a custom serialization
functionality.