gh-24: implement removing embeddings from an index
After initial full scan of embeddings for the ScyllaDB there is a need
to support possible removal of embeddings from an index. This patch
introduces this information by refactoring DbEmbeddings::embeddings to
the optional field - None means there is a removal request. Both add
and removal requests have specific timestamps for coordinating merging
of embeddings.