gh-24: update scylla rust driver
Using CDC ScyllaDB functionality needs scylla-cdc and async-trait
crates. As vector type support is not merged with scylla driver there is
a need for integration branch on a separate repository. This patch uses
ewienik for a scylla driver crate. Eventually vector support will be
merged and then scylla and scylla-cdc crates version should be updated.