gh-24: implement reading new embeddings from the CDC
This patch implements support for reading CDC in a concurrent task and
push outcome data into copy of the mpsc channel. The task is started
before initial full scan, to be able to catch all data written into the
table after the fullscan finish. There is a possibility that some data
will be duplicated (same in fullscan and in CDC) but merging will be
solved by using timestamps. Timestamps from the CDC functionality in the
scylla-cdc crate are compatible with UUIDv1 timestamp format, which is
hunderds of nanosecods since gregorian epoch.

Fixes: #24